prima di semplificare il factoy method

+++ MultiWeatherApi/DarkSky/Model/Forecast.cs
@@ -19,62 +19,48 @@ namespace MultiWeatherApi.DarkSky.Model {
             set => _coordinates = value;
         }
 
-        /// <summary>
-        ///     the IANA time zone name for this location.
-        /// </summary>
+        /// <summary>the IANA time zone name for this location</summary>
         [JsonProperty("timezone")]
         public string TimeZone { get; set; }
 
-        /// <summary>
-        ///     the time zone offset, in hours from GMT.
-        /// </summary>
+        /// <summary>the time zone offset, in hours from GMT.</summary>
         [JsonProperty("offset")]
-        public double TimeZoneOffset { get; set; }
+        public float TimeZoneOffset { get; set; }
 
-        /// <summary>
-        ///     the current conditions at the requested location.
-        /// </summary>
+        /// <summary>the current conditions at the requested location.</summary>
         [JsonProperty("currently")]
         public DataPoint Currently { get; set; }
 
-        /// <summary>
-        ///     the minute-by-minute conditions for the next hour.
-        /// </summary>
+        /// <summary>the minute-by-minute conditions for the next hour.</summary>
         [JsonProperty("minutely")]
         public ForecastDetails Minutely { get; set; }
 
-        /// <summary>
-        ///     the hour-by-hour conditions for the next two days.
-        /// </summary>
+        /// <summary>the hour-by-hour conditions for the next two days.</summary>
         [JsonProperty("hourly")]
         public ForecastDetails Hourly { get; set; }
 
-        /// <summary>
-        ///     the daily conditions for the next week.
-        /// </summary>
+        /// <summary>the daily conditions for the next week.</summary>
         [JsonProperty("daily")]
         public ForecastDetails Daily { get; set; }
 
-        /// <summary>
-        ///     the metadata (flags) associated with this forecast.
-        /// </summary>
+        /// <summary>the metadata (flags) associated with this forecast.</summary>
         [JsonProperty("flags")]
         public Flags Flags { get; set; }
 
-        /// <summary>
-        ///     any weather alerts related to this location.
-        /// </summary>
+        /// <summary>Any weather alerts related to this location.</summary>
         [JsonProperty("alerts")]
         public IList<Alert> Alerts { get; set; }
 
         #region Internals
 
+        /// <summary>the latitude of this forecast.</summary>
         [JsonProperty("latitude")]
         internal double Latitude {
             //get => _coordinates.Latitude;
             set => _coordinates.Latitude = value;
         }
 
+        /// <summary>the longitude of this forecast.</summary>
         [JsonProperty("longitude")]
         internal double Longitude {
             //get => _coordinates.Longitude;
